requirements.txt
memori
openai
google-genai
SQLAlchemy
flask
python-dotenv

